Made Time Conversion to Float Explicit

+++ Engine/src/Oyl3D/Platform/Windows/Win32Platform.cpp
@@ -21,7 +21,7 @@ float Win32Platform::getTimeImpl() {
 	}();
 
 	QueryPerformanceCounter(&li);
-	return double(li.QuadPart - timeStart) / frequency;
+	return static_cast<float>(double(li.QuadPart - timeStart) / frequency);
 }
 
 }
